pyproject.toml
[project]
name = "clubtwitshows"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "lxml>=6.0.1",
    "pyside6>=6.9.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
run-requirements = "install_requirements:main"
